requirements.txt
Django>=3.0,<4.0
djangorestframework
djangorestframework_simplejwt
psycopg2-binary>=2.8
celery==4.4.1
redis==3.4.1